server-python/requirements.txt
numpy
pandas
plotly
scikit-learn
tensorflow
keras
Flask
dash
dash-html-components
dash-core-components
streamlit
python-binance
python-dotenv
django
djangorestframework
path